Add Access Token API
Adds the ability to create access tokens from the current process and
also on behalf of a user (given their credentials). In addition, tokens
can be further restricted using the token.CreateRestrictedToken
function. This can be used to create processes with more limited
privileges.

- Get CurrentProcessToken
- RunAs uses the ImpersonateLoggedOnUser API to run a function in the context
  of a given user's access token
- Get Environment from token
- Add Access Mask Constants

This depends on two more modules which have been added to go.mod

- github.com/pkg/errors : smart error wrapping with the ability to stack
traces
- golang.org/x/sys/windows : additional helper functions for calling the
win32 api